packaged-agent/pyproject.toml
[project]
name = "packaged-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "umair86450", email = "2001muhammadumair@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.10",
    "python-dotenv>=1.1.1",
]

[project.scripts]
packaged-agent = "packaged_agent:main"
agent = "packaged_agent.main:start_study_assistant"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
